refactor: split metrics/sinks/clickhouse.py by table family (CHAOS-1541)
Convert the 1996-line monolithic ClickHouseMetricsSink into a package:

  src/dev_health_ops/metrics/sinks/clickhouse/
  ├── __init__.py     # re-exports ClickHouseMetricsSink (stable public API)
  ├── _insert.py      # DEFAULT_BATCH_SIZE, _chunked, _dt_to_clickhouse_datetime
  ├── core.py         # ClickHouseCore: __init__, close, schema, _insert_rows, helpers
  ├── ci.py           # CIMixin: cicd/deploy/incident/testops/feature-flags/telemetry
  ├── dora.py         # DoraMixin: dora_metrics + benchmark/comparison tables
  ├── wellbeing.py    # WellbeingMixin: user_metrics, quality_drag, pipeline_stability
  ├── investment.py   # InvestmentMixin: investment*/work_unit*/explanation tables
  └── work_graph.py   # WorkGraphMixin: repo/file/commit/work-graph/items/forecasts

Public API preserved:
  from dev_health_ops.metrics.sinks.clickhouse import ClickHouseMetricsSink

All 52 write_* methods present. MRO fix: mixins before ClickHouseCore so
concrete implementations override abstract methods from BaseMetricsSink.
Zero new SQL, column, or behavior changes — pure structural split.

Before: 1996 lines (1 file)
After:  2238 lines across 8 files (avg 280 lines/file)
